Initial code push for cfDBManager
This push includes a mindmap of the initial development stage of
cfDBManager, which is slated to include backup/restore and import/export
functions in various formats.

The code is being designed to be used in an MVC style framework, though
a framework has not yet been decided by the team. Some models, services
and preliminary code has been constructed and is included in this update
along with structure and layout. Ultimately this code plans to utilize
CF Gateway functionality in an effort to provide for large data
import/export/backup/restore functions without timing out or bogging
down the server.